+++ foliant/preprocessors/includes.py
@@ -589,7 +589,8 @@ class Preprocessor(BasePreprocessor):
         self,
         markdown_file_path: Path,
         content: str,
-        project_root_path: Path or None = None
+        project_root_path: Path or None = None,
+        sethead: int or None = None
     ) -> str:
         '''Replace all include statements with the respective file contents.
 
@@ -597,6 +598,8 @@ class Preprocessor(BasePreprocessor):
         :param content: Markdown content
         :param project_root_path: Path to the “root” directory of Foliant project
             that the currently processed Markdown file belongs to
+        :param sethead: Level of the topmost heading in the content,
+            it may be set when the method is called recursively
 
         :returns: Markdown content with resolved includes
         '''
@@ -618,7 +621,6 @@ class Preprocessor(BasePreprocessor):
             include_statement = self.pattern.fullmatch(content_part)
 
             if include_statement:
-                # Reset current project root before each new include statement
                 current_project_root_path = project_root_path
 
                 body = self._tag_body_pattern.match(include_statement.group('body').strip())
@@ -628,6 +630,22 @@ class Preprocessor(BasePreprocessor):
                     f'Processing include statement; body: {body}, options: {options}, ' +
                     f'current project root path: {current_project_root_path}'
                 )
+
+                current_sethead = sethead
+
+                self.logger.debug(
+                    f'Current sethead: {current_sethead}, ' +
+                    f'user-specified sethead: {options.get("sethead")}'
+                )
+
+                if options.get('sethead'):
+                    if current_sethead:
+                        current_sethead += options['sethead'] - 1
+
+                    else:
+                        current_sethead = options['sethead']
+
+                    self.logger.debug(f'Set new current sethead: {current_sethead}')
 
                 # If the tag body is not empty, the legacy syntax is expected:
                 #
@@ -700,7 +718,7 @@ class Preprocessor(BasePreprocessor):
                             project_root_path=current_project_root_path,
                             from_heading=body.group('from_heading'),
                             to_heading=body.group('to_heading'),
-                            sethead=options.get('sethead'),
+                            sethead=current_sethead,
                             nohead=options.get('nohead')
                         )
 
@@ -728,7 +746,7 @@ class Preprocessor(BasePreprocessor):
                             project_root_path=current_project_root_path,
                             from_heading=body.group('from_heading'),
                             to_heading=body.group('to_heading'),
-                            sethead=options.get('sethead'),
+                            sethead=current_sethead,
                             nohead=options.get('nohead')
                         )
 
@@ -759,7 +777,7 @@ class Preprocessor(BasePreprocessor):
                             to_heading=options.get('to_heading'),
                             from_id=options.get('from_id'),
                             to_id=options.get('to_id'),
-                            sethead=options.get('sethead'),
+                            sethead=current_sethead,
                             nohead=options.get('nohead')
                         )
 
@@ -784,7 +802,7 @@ class Preprocessor(BasePreprocessor):
                             to_heading=options.get('to_heading'),
                             from_id=options.get('from_id'),
                             to_id=options.get('to_id'),
-                            sethead=options.get('sethead'),
+                            sethead=current_sethead,
                             nohead=options.get('nohead')
                         )
 
@@ -801,7 +819,8 @@ class Preprocessor(BasePreprocessor):
                     processed_content_part = self.process_includes(
                         included_file_path,
                         processed_content_part,
-                        current_project_root_path
+                        current_project_root_path,
+                        current_sethead
                     )
 
                 if options.get('inline'):